worker/requirements.txt
temporalio
httpx
psycopg2-binary
redis
pytest
reportlab
pyyaml
asyncpg>=0.29.0
zstandard>=0.21.0
scikit-learn>=1.3.0
# Pydantic Settings — centralized config
pydantic-settings>=2.0
# OpenTelemetry traces + logs (optional — gracefully degrades if missing)
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp-proto-http>=1.27.0
# Optional data plane (Redpanda / analytics)
kafka-python>=2.0.2
duckdb>=1.0.0
pyarrow>=14.0.0
# torch is NOT listed here — install separately (2GB+):
#   pip install torch --index-url https://download.pytorch.org/whl/cu118
